feat: add support for craft 3.7.27
