+++ README.md
@@ -22,16 +22,6 @@ To install the plugin, follow these instructions.
 
 3. In the Control Panel, go to Settings → Plugins and click the “Install” button for FAB Permissions.
 
-4. **Important** - Override the Craft Fields Service with the FAB Permissions Fields Service in `config/app.web.php`:
-
-		return [
-			'components' => [
-				'fields' => [
-					'class' => 'thejoshsmith\fabpermissions\services\Fields'
-				]
-			]
-		]
-
 ## FAB Permissions Overview
 
 This plugin allows you to restrict access to certain user groups on a per tab or field basis.
